requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
backoff==2.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
docopt==0.6.2
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
haystack-ai==2.7.0
haystack-experimental==0.6.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lazy_imports==0.4.0
MarkupSafe==3.0.2
mistral-haystack==0.0.2
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openai==1.64.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pipreqs==0.4.13
posthog==3.16.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wikipedia==1.4.0
yarg==0.1.10
